+++ lib/msal-browser/src/request/AuthorizationCodeRequest.ts
@@ -5,7 +5,7 @@
 
 import { CommonAuthorizationCodeRequest } from "@azure/msal-common";
 
-export type AuthorizationCodeRequest = Partial<Omit<CommonAuthorizationCodeRequest, "code"|"enableSpaAuthorizationCode">> & {
+export type AuthorizationCodeRequest = Partial<Omit<CommonAuthorizationCodeRequest, "code"|"enableSpaAuthorizationCode"|"requestedClaimsHash">> & {
     code: string;
     cloudGraphHostName?: string;
     msGraphHost?: string;