+++ lib/msal-browser/test/app/PublicClientApplication.spec.ts
@@ -5,7 +5,7 @@
 
 import sinon from "sinon";
 import { PublicClientApplication } from "../../src/app/PublicClientApplication";
-import { TEST_CONFIG, TEST_URIS, TEST_TOKENS, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES, RANDOM_TEST_GUID, testNavUrl, testLogoutUrl, TEST_STATE_VALUES, TEST_HASHES, DEFAULT_TENANT_DISCOVERY_RESPONSE, DEFAULT_OPENID_CONFIG_RESPONSE, testNavUrlNoRequest, TEST_SSH_VALUES } from "../utils/StringConstants";
+import { TEST_CONFIG, TEST_URIS, TEST_TOKENS, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES, RANDOM_TEST_GUID, testNavUrl, testLogoutUrl, TEST_STATE_VALUES, TEST_HASHES, DEFAULT_TENANT_DISCOVERY_RESPONSE, DEFAULT_OPENID_CONFIG_RESPONSE, testNavUrlNoRequest, TEST_SSH_VALUES, TEST_CRYPTO_VALUES } from "../utils/StringConstants";
 import { ServerError, Constants, AccountInfo, TokenClaims, AuthenticationResult, CommonAuthorizationUrlRequest, AuthorizationCodeClient, ResponseMode, AccountEntity, ProtocolUtils, AuthenticationScheme, RefreshTokenClient, Logger, ServerTelemetryEntity, CommonSilentFlowRequest, LogLevel, CommonAuthorizationCodeRequest } from "@azure/msal-common";
 import { ApiId, InteractionType, WrapperSKU, TemporaryCacheKeys, BrowserConstants, BrowserCacheLocation } from "../../src/utils/BrowserConstants";
 import { CryptoOps } from "../../src/crypto/CryptoOps";
@@ -977,12 +977,14 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 tokenType: AuthenticationScheme.BEARER
             };
             sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+            sinon.stub(CryptoOps.prototype, "hashString").resolves(TEST_CRYPTO_VALUES.TEST_SHA256_HASH);
             const silentATStub = sinon.stub(RefreshTokenClient.prototype, "acquireTokenByRefreshToken").resolves(testTokenResponse);
             const tokenRequest: CommonSilentFlowRequest = {
                 scopes: ["User.Read"],
                 account: testAccount,
                 authority: TEST_CONFIG.validAuthority,
                 authenticationScheme: AuthenticationScheme.BEARER,
+                claims: JSON.stringify({claim: "claim"}),
                 correlationId: TEST_CONFIG.CORRELATION_ID,
                 forceRefresh: false
             };
@@ -991,8 +993,11 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 scopes: ["User.Read"],
                 authority: `${Constants.DEFAULT_AUTHORITY}`,
                 correlationId: RANDOM_TEST_GUID,
+                claims: JSON.stringify({claim: "claim"}),
+                requestedClaimsHash: TEST_CRYPTO_VALUES.TEST_SHA256_HASH,
                 forceRefresh: false
             };
+
             const silentRequest1 = pca.acquireTokenSilent(tokenRequest);
             const silentRequest2 = pca.acquireTokenSilent(tokenRequest);
             const silentRequest3 = pca.acquireTokenSilent(tokenRequest);
@@ -1048,6 +1053,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 tokenType: AuthenticationScheme.BEARER
             };
             sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+            sinon.stub(CryptoOps.prototype, "hashString").resolves(TEST_CRYPTO_VALUES.TEST_SHA256_HASH);
             const silentATStub = sinon.stub(RefreshTokenClient.prototype, "acquireTokenByRefreshToken").resolves(testTokenResponse);
             // Beaerer requests
             const tokenRequest1: CommonSilentFlowRequest = {
@@ -1158,6 +1164,48 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 forceRefresh: false
             };
 
+            // Requests with claims
+            const claimsRequest1: CommonSilentFlowRequest = {
+                scopes: ["User.Read"],
+                account: testAccount,
+                authority: TEST_CONFIG.validAuthority,
+                authenticationScheme: AuthenticationScheme.BEARER,
+                claims: JSON.stringify({ claim1: "claim1"}),
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                forceRefresh: false
+            }
+
+            const claimsRequest2: CommonSilentFlowRequest = {
+                scopes: ["User.Read"],
+                account: testAccount,
+                authority: TEST_CONFIG.validAuthority,
+                authenticationScheme: AuthenticationScheme.BEARER,
+                claims: JSON.stringify({ claim2: "claim2"}),
+                requestedClaimsHash: TEST_CRYPTO_VALUES.TEST_SHA256_HASH,
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                forceRefresh: false
+            }
+
+            const expectedClaimsRequest1: CommonSilentFlowRequest = {
+                ...claimsRequest1,
+                scopes: ["User.Read"],
+                authority: `${Constants.DEFAULT_AUTHORITY}`,
+                correlationId: RANDOM_TEST_GUID,
+                claims: JSON.stringify({ claim1: "claim1"}),
+                requestedClaimsHash: TEST_CRYPTO_VALUES.TEST_SHA256_HASH,
+                forceRefresh: false
+            };
+
+            const expectedClaimsRequest2: CommonSilentFlowRequest = {
+                ...claimsRequest2,
+                scopes: ["User.Read"],
+                authority: `${Constants.DEFAULT_AUTHORITY}`,
+                correlationId: RANDOM_TEST_GUID,
+                claims: JSON.stringify({ claim2: "claim2"}),
+                requestedClaimsHash: TEST_CRYPTO_VALUES.TEST_SHA256_HASH,
+                forceRefresh: false
+            };
+
             const silentRequest1 = pca.acquireTokenSilent(tokenRequest1);
             const silentRequest2 = pca.acquireTokenSilent(tokenRequest1);
             const silentRequest3 = pca.acquireTokenSilent(tokenRequest2);
@@ -1167,7 +1215,15 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             const sshCertSilentRequest1 = pca.acquireTokenSilent(sshCertRequest1);
             const sshCertSilentRequest2 = pca.acquireTokenSilent(sshCertRequest1);
             const sshCertSilentRequest3 = pca.acquireTokenSilent(sshCertRequest2);
-            await Promise.all([silentRequest1, silentRequest2, silentRequest3, popSilentRequest1, popSilentRequest2, popSilentRequest3, sshCertSilentRequest1, sshCertSilentRequest2, sshCertSilentRequest3]);
+            const claimsSilentRequest1 = pca.acquireTokenSilent(claimsRequest1);
+            const claimsSilentRequest2 = pca.acquireTokenSilent(claimsRequest1);
+            const claimsSilentRequest3 = pca.acquireTokenSilent(claimsRequest2);
+            await Promise.all([
+                silentRequest1, silentRequest2, silentRequest3,
+                popSilentRequest1, popSilentRequest2, popSilentRequest3,
+                sshCertSilentRequest1, sshCertSilentRequest2, sshCertSilentRequest3,
+                claimsSilentRequest1, claimsSilentRequest2, claimsSilentRequest3
+            ]);
 
             expect(silentATStub.calledWith(expectedTokenRequest1)).toBeTruthy();
             expect(silentATStub.calledWith(expectedTokenRequest2)).toBeTruthy();
@@ -1175,7 +1231,9 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             expect(silentATStub.calledWith(expectedPopTokenRequest2)).toBeTruthy();
             expect(silentATStub.calledWith(expectedSshCertificateRequest1)).toBeTruthy();
             expect(silentATStub.calledWith(expectedSshCertificateRequest2)).toBeTruthy();
-            expect(silentATStub.callCount).toEqual(6);
+            expect(silentATStub.calledWith(expectedClaimsRequest1)).toBeTruthy();
+            expect(silentATStub.calledWith(expectedClaimsRequest2)).toBeTruthy();
+            expect(silentATStub.callCount).toEqual(8);
         });
 
         it("throws error that SilentFlowClient.acquireToken() throws", async () => {